pyproject.toml
[project]
name = "everybody-codes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "LeoPedri", email = "leonardopedri1.0@gmail.com" }
]
requires-python = ">=3.13"
dependencies = []

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
cli = [
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer[all]>=0.20.0",
]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.4",
    "ty>=0.0.1a25",
]

[project.scripts]
ec = "everybody_codes.cli:main"
